Login in using the provided user name
